fix(create-changelog): Validate the existence of commits at version pattern match
Matching a version pattern in the first commit indicates the changes are from the second latest commit up to the second commit to match the pattern. Thus, the action should only stop at a patterns when the pattern doesn't happen at the first commit. This fixes a bug where the first commit was being considered when the description matched the pattern.