feat(create-changelog): thank non-regular contributors
This feature thanks non-regular contributors who created commits or opened issues related to these commits. Any user who is the repo admin, owner, or affiliated to the repo owner is considered regular. Any user with less 10% of commits, or less than 3 commits, or not in the 20% top contributors is considered non-regular. Everyone else is considered regular.