feat(flamegraph): Report includes Project Symbols section
The Symbols section tends to be dominated by the STL because they tend to be included multiple times. It's useful to know about the bottlenecks in the STL, since removing unnecessary symbols provide the most gains. But because we have no control over the STL, we might get to a point where there is no additional benefits in excluding symbols. STL symbols also tend to be less representative in cases where the user is already expected to include that header. At this point, it's more useful to have a report about symbols outside the STL, which is where we have more control to improve build times.